refactor: ♻️ reorganización de 'features/login' y 'app/(pages)/auth/login'
- Se separan los archivos de la feature en carpetas por propósito.
- Referencias actualizadas en auth/login/page.tsx para reflejar los cambios.
