+++ src/features/login/types/LoginCredentials.ts
@@ -1,0 +1,4 @@
+export interface LoginCredentials {
+  email: string;
+  password: string;
+}
