+++ src/features/login/types/LoginFormData.ts
@@ -1,0 +1,4 @@
+export interface LoginFormData {
+  email: string;
+  password: string;
+}
