+++ src/features/login/types/ToastData.ts
@@ -1,0 +1,4 @@
+export interface ToastData {
+  message: string;
+  type: "success" | "error" | "";
+}
